moved functions out of main.py
moved the functions in main.py out to seperate file to start work on curses and improve overall functionality, also to make commits more
readable such as changing the functions or how the program works as a whole
