+++ main.py
@@ -1,6 +1,5 @@
-import base64
-import random
-import string
+import curses
+import cipher
 
 # this code takes an input then encodes the code with the Vigenère cipher, which takes a password to encode or decode
 # the note, the colin defines the start, if there is a space directly after the colin, its part of the code
@@ -21,74 +20,10 @@ key = input()
 print('enter text to change: ')
 user = input()
 
-def encode(a, b):
-    x = len(b)
-    y = len(a) - 1
-    z = -1
-    realkey = []
-    for i in range(x):
-        z += 1
-        realkey.append(a[z])
-        if z >= y:
-            z = -1
-    letters = string.ascii_letters
-    digits = string.digits
-    symbols = string.punctuation
-    chars = list(letters + digits + symbols)
-    chars.append(' ')
-    ans = []
-    for i in range(x):
-        cypher = chars.index(realkey[i])
-        let = chars.index(b[i])
-        a = let + cypher
-        if a > len(chars):
-            a -= len(chars)
-        ans.append(chars[a])
-    return ''.join(ans)
-
-
-def decode(a, b):
-    x = len(b)
-    y = len(a) - 1
-    z = -1
-    realkey = []
-    encry = []
-    for i in range(x):
-        z += 1
-        realkey.append(a[z])
-        if z >= y:
-            z = -1
-    letters = string.ascii_letters
-    digits = string.digits
-    symbols = string.punctuation
-    chars = list(letters + digits + symbols)
-    chars.append(' ')
-    ans = []
-    for i in range(x):
-        cypher = chars.index(realkey[i])
-        let = chars.index(b[i])
-        a = let - cypher
-        if a < 0:
-            a += len(chars)
-        ans.append(chars[a])
-    return ''.join(ans)
-
-def passwordgen(legnth):
-    letters = string.ascii_letters
-    digits = string.digits
-    symbols = string.punctuation
-    chars = letters + digits + symbols
-
-    password = ''
-    for i in range(legnth):
-        password += random.choice(chars)
-    return(password)
-
-
 if key == '':
-    key = passwordgen(16)
+    key = cipher.passwordgen(16)
 if endecpt == 0:
-    print('encrypted string:' + str(encode(key, user)))
+    print('encrypted string:' + str(cipher.encode(key, user)))
     print('key: ' + str(key))
 if endecpt == 1:
-    print('message: ' + str(decode(key, user)))
+    print('message: ' + str(cipher.decode(key, user)))
